+++ public/x/service/service.go
@@ -1,6 +1,8 @@
 package service
 
 import (
+	"context"
+
 	"github.com/Jeffail/benthos/v3/lib/service"
 )
 
@@ -8,8 +10,14 @@ import (
 // file path(s) and execute subcommands for linting configs, testing configs,
 // etc. This is how a standard distribution of Benthos operates.
 //
-// This call blocks until either the service shuts down or a termination signal
-// is received.
-func RunCLI() {
-	service.Run()
+// This call blocks until either:
+// - The service shuts down gracefully due to the inputs closing
+// - A termination signal is received
+// - The provided context is cancelled
+//
+// This function must only be called once during the entire lifecycle of your
+// program, as it interacts with singleton state. In order to manage multiple
+// Benthos stream lifecycles in a program use the StreamBuilder API instead.
+func RunCLI(ctx context.Context) {
+	service.RunWithOpts(service.OptUseContext(ctx))
 }